requirements.txt
wheel
Pillow
gunicorn
psycopg2-binary
whitenoise
django-dotenv
argon2-cffi
argon2-cffi-bindings
django
django-htmx
qrcode==7.3.1
django-storages
boto3
celery
django-allauth
